+++ plugins/scaffolder-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-scaffolder-backend
 
+## 0.9.5
+
+### Patch Changes
+
+- 802b41b65: Allow custom directory to be specified for GitHub publish action
+- Updated dependencies [97b60de98]
+- Updated dependencies [98dd5da71]
+- Updated dependencies [b779b5fee]
+  - @backstage/catalog-model@0.7.6
+  - @backstage/backend-common@0.6.2
+
 ## 0.9.4
 
 ### Patch Changes